Rename Disclaimer to SecurityFeedbackModal
