+++ src/routes/safe/components/Apps/index.tsx
@@ -3,7 +3,7 @@ import { useHistory } from 'react-router-dom'
 import { useSafeAppUrl } from 'src/logic/hooks/useSafeAppUrl'
 import AppFrame from 'src/routes/safe/components/Apps/components/AppFrame'
 import AppsList from 'src/routes/safe/components/Apps/components/AppsList'
-import SafeAppsDisclaimer from './components/Disclaimer'
+import SafeAppsDisclaimer from './components/SecurityFeedbackModal'
 import SafeAppsErrorBoundary from './components/SafeAppsErrorBoundary'
 import SafeAppsLoadError from './components/SafeAppsLoadError'
 import { useSecurityFeedbackModal } from './hooks/useSecurityFeedbackModal'
